Fix relationship loading for models with dashes
Previously when we encountered a relationship with a model that included
a dash in the name we would mangle to the relationship name in the
process of trying to find the relationship config. This caused us to
fail to find the actual data for the relationship deeper down in the
stack when we call `relatedItemsFor`.

This change ensures that the original relationship name is kept
unmangled so that we can use it later to look up the data.

As a side note the current logic to try and find the relationship
configuration is a little brittle with regard to the way that it
actually finds the fields that the included data maps to. We could
probably go further here and just scan all of the fields to see if any
of them are relationship configs and then populate whatever we find
(where data has been included).