Throw an Error if id is nil
Fail fast if the caller tries to use `null` or `undefined` as the id.
Nothing good will come of it so you'd better know about it as soon as
possible.

Note that it's still possible to use the strings "null" or "undefined"
as the id which will construct the URLs you'd expect it to.